Move language definitions to separate CFG files
