+++ src/parser-loader/bun.ts
@@ -6,7 +6,9 @@ export async function initializeParser(
 ): Promise<{ parser: Parser; language: TreeSitterLanguage }> {
   await Parser.init();
 
-  const parserLanguage = await TreeSitterLanguage.load(languageDefinitions[language].wasmPath);
+  const parserLanguage = await TreeSitterLanguage.load(
+    languageDefinitions[language].wasmPath,
+  );
   const parser = new Parser();
   parser.setLanguage(parserLanguage);
   return { parser, language: parserLanguage };
